[PR requested change] Nest consumer/producer config props under pulsar props
